// playground-solid/src/App/PageComponents/StressTest/StressText.types.ts
import type { JSX } from "solid-js";

import type { AccessorProps } from "@thewaver/ss-components-solid";

export type StressTestDefs = {
    count: number;
    cols: number;
    gap: number;
};

export type StressTestProps = AccessorProps<{
    configs: StressTestDefs[];
    onShowModal?: () => void;
    onHideModal?: () => void;
    renderLabel: (getConfigIndex: () => number) => JSX.Element;
    renderItem: (getConfigIndex: () => number, getItemIndex: () => number) => JSX.Element;
}>;